requirements.txt
streamlit
tensorflow-cpu
pillow
scikit-learn
pandas
nltk
numpy<2.0.0